add pageable

+++ src/main/java/com/example/fleet/services/TaxiService.java
@@ -3,37 +3,36 @@ package com.example.fleet.services;
 import com.example.fleet.models.TaxiModel;
 import com.example.fleet.repositories.TaxiRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-
-import java.util.ArrayList;
-import java.util.Optional;
 
 @Service
 public class TaxiService {
     @Autowired
     TaxiRepository taxiRepository;
 
-    public ArrayList<TaxiModel> getTaxis(){
-        return (ArrayList<TaxiModel>) taxiRepository.findAll();
-    }
-    public TaxiModel saveTaxi(TaxiModel taxi){
-        return taxiRepository.save(taxi);
-    }
-    // Debe regresar uno solo cambio el array???
-    public ArrayList<TaxiModel> getByPlate(String plate){
-        return taxiRepository.findByPlate(plate);
+    public Page<TaxiModel> getTaxis(Pageable pageable) {
+        return taxiRepository.findAll(pageable);
     }
 
-    public Optional<TaxiModel> getById(Integer id){
-        return taxiRepository.findById(id);
+    public TaxiModel saveTaxi(TaxiModel taxi) {
+        return taxiRepository.save(taxi);
+    }
+
+    public TaxiModel getById(Integer id) {
+        return taxiRepository.findById(id).orElse(null);
+    }
+
+    public Page<TaxiModel> getByPlate(String plate, Pageable pageable) {
+        return taxiRepository.findByPlate(plate, pageable);
     }
 
     public boolean deleteTaxi(Integer id) {
-        try{
+        if (taxiRepository.existsById(id)) {
             taxiRepository.deleteById(id);
             return true;
-        }catch(Exception err){
-            return false;
         }
+        return false;
     }
 }
